pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "craftax"
version = "1.2.0"
authors = [
  { name="Michael Matthews", email="michael.matthews@eng.ox.ac.uk" },
]
description = "An open-world environment for training RL agents"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "jax",
    "distrax",
    "optax",
    "flax",
    "numpy",
    "black",
    "pre-commit",
    "argparse",
    "wandb",
    "orbax-checkpoint",
    "pygame",
    "gymnax",
    "chex",
    "matplotlib",
    "imageio",
    "dm-haiku"
]

[project.scripts]
play_craftax = "craftax.craftax.play_craftax:entry_point"
play_craftax_classic = "craftax.craftax_classic.play_craftax_classic:main"

[project.urls]
Homepage = "https://github.com/MichaelTMatthews/Craftax"
Issues = "https://github.com/MichaelTMatthews/Craftax/issues"

[tool.setuptools.packages.find]
include=["craftax*"]